EOS-14652 S3 SW Update - Configs Merge

+++ rpms/s3/s3rpm.spec
@@ -122,6 +122,20 @@ S3 server provides S3 REST API interface support for Motr object storage.
 %prep
 %setup -n %{name}-%{version}-%{_s3_git_ver}
 
+%pre
+if [ -f /opt/seagate/cortx/s3/conf/s3config.yaml.sample ]; then
+    cp -f /opt/seagate/cortx/s3/conf/s3config.yaml.sample /tmp/s3config.yaml.sample.old
+fi
+if [ -f /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml.sample ]; then
+    cp -f /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml.sample /tmp/config.yaml.sample.old
+fi
+if [ -f /opt/seagate/cortx/auth/resources/keystore.properties.sample ]; then
+    cp -f /opt/seagate/cortx/auth/resources/keystore.properties.sample /tmp/keystore.properties.sample.old
+fi
+if [ -f /opt/seagate/cortx/auth/resources/authserver.properties.sample ]; then
+    cp -f /opt/seagate/cortx/auth/resources/authserver.properties.sample /tmp/authserver.properties.sample.old
+fi
+
 %build
 %if %{with cortx_motr}
 ./rebuildall.sh --no-check-code --no-install
@@ -210,6 +224,7 @@ rm -rf %{buildroot}
 /opt/seagate/cortx/auth/AuthServer-1.0-0.jar
 /opt/seagate/cortx/auth/AuthPassEncryptCLI-1.0-0.jar
 /opt/seagate/cortx/auth/startauth.sh
+/opt/seagate/cortx/auth/scripts/swupdate/merge.sh
 /opt/seagate/cortx/auth/scripts/enc_ldap_passwd_in_cfg.sh
 /opt/seagate/cortx/auth/scripts/change_ldap_passwd.ldif
 /opt/seagate/cortx/auth/scripts/s3authserver.jks_template
@@ -305,6 +320,11 @@ rm -rf %{buildroot}
 /opt/seagate/cortx/s3/conf/setup.yaml
 /opt/seagate/cortx/s3/s3datarecovery/s3_data_recovery.sh
 /opt/seagate/cortx/datarecovery/orchastrator.sh
+/opt/seagate/cortx/auth/resources/authserver_safe_attributes.yaml
+/opt/seagate/cortx/auth/resources/keystore_safe_attributes.yaml
+/opt/seagate/cortx/s3/conf/s3config_safe_attributes.yaml
+/opt/seagate/cortx/s3/s3backgrounddelete/s3backgrounddelete_safe_attributes.yaml
+
 %attr(755, root, root) /opt/seagate/cortx/s3/bin/s3_setup
 %attr(755, root, root) /opt/seagate/cortx/s3/s3backgrounddelete/s3backgroundconsumer
 %attr(755, root, root) /opt/seagate/cortx/s3/s3backgrounddelete/s3backgroundproducer
@@ -336,14 +356,32 @@ rm -rf %{buildroot}
 %exclude /opt/seagate/cortx/s3/reset/precheck.pyo
 
 %post
-[ -f /opt/seagate/cortx/s3/conf/s3config.yaml ] || 
+if [ -f /opt/seagate/cortx/s3/conf/s3config.yaml ]; then
+sh /opt/seagate/cortx/auth/scripts/merge.sh /opt/seagate/cortx/s3/conf/s3config.yaml /tmp/s3config.yaml.sample.old /opt/seagate/cortx/s3/conf/s3config.yaml.sample /opt/seagate/cortx/s3/conf/s3config_safe_attributes.yaml yaml
+else
     cp /opt/seagate/cortx/s3/conf/s3config.yaml.sample /opt/seagate/cortx/s3/conf/s3config.yaml
-[ -f /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml ] ||
+fi
+if [ -f /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml ]; then
+sh /opt/seagate/cortx/auth/scripts/merge.sh /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml /tmp/config.yaml.sample.old /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml.sample /opt/seagate/cortx/s3/s3backgrounddelete/s3backgrounddelete_safe_attributes.yaml yaml
+else
     cp /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml.sample /opt/seagate/cortx/s3/s3backgrounddelete/config.yaml
-[ -f /opt/seagate/cortx/auth/resources/authserver.properties ] ||
+fi
+if [ -f /opt/seagate/cortx/auth/resources/authserver.properties ]; then
+sh /opt/seagate/cortx/auth/scripts/merge.sh /opt/seagate/cortx/auth/resources/authserver.properties /tmp/authserver.properties.sample.old /opt/seagate/cortx/auth/resources/authserver.properties.sample /opt/seagate/cortx/auth/resources/authserver_safe_attributes.yaml properties
+else
     cp /opt/seagate/cortx/auth/resources/authserver.properties.sample /opt/seagate/cortx/auth/resources/authserver.properties
-[ -f /opt/seagate/cortx/auth/resources/keystore.properties ] ||
+fi
+if [ -f /opt/seagate/cortx/auth/resources/keystore.properties ]; then
+sh /opt/seagate/cortx/auth/scripts/merge.sh /opt/seagate/cortx/auth/resources/keystore.properties /tmp/keystore.properties.sample.old /opt/seagate/cortx/auth/resources/keystore.properties.sample /opt/seagate/cortx/auth/resources/keystore_safe_attributes.yaml properties
+else
     cp /opt/seagate/cortx/auth/resources/keystore.properties.sample /opt/seagate/cortx/auth/resources/keystore.properties
+fi
+
+rm -f /tmp/s3config.yaml.sample.old
+rm -f /tmp/config.yaml.sample.old
+rm -f /tmp/authserver.properties.sample.old
+rm -f /tmp/keystore.properties.sample.old
+
 systemctl daemon-reload
 systemctl enable s3authserver
 systemctl restart rsyslog
